[demo release] API modify; modify npm scripts
